+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Use `JTextPane` instead of `JTextArea` to avoid error layout when component is RTL (issues with layout)
 - Drawer now use shadow border
 - Update from layout
+- Quick search update layout and add new favorite option (PR #17)
 
 ## [2.0.0] - 2024-10-20