+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [2.1.0]-SNAPSHOT
+## [2.1.0] - 2024-11-07
 
 ### New features and improvements